+++ docs/history.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+`2.0.0`_ (17 Oct 2016)
+----------------------
+- Breaking API change for Client.get_item to allow for return values for ConsumedCapacity
+- Implement Client.delete_item, Client.update_item, Client.query, Client.scan
+- Improved parameter validation
+
 `1.1.0`_ (12 Oct 2016)
 ----------------------
 - Remove the service tag in InfluxDB
@@ -24,7 +30,8 @@ Release History
 `Next Release`_
 ---------------
 
-.. _Next Release: https://github.com/sprockets/sprockets_dynamodb/compare/1.1.0...master
+.. _Next Release: https://github.com/sprockets/sprockets_dynamodb/compare/2.0.0...master
+.. _2.0.0: https://github.com/sprockets/sprockets-dynamodb/compare/1.1.0...2.0.0
 .. _1.1.0: https://github.com/sprockets/sprockets-dynamodb/compare/1.0.2...1.1.0
 .. _1.0.2: https://github.com/sprockets/sprockets-dynamodb/compare/1.0.1...1.0.2
 .. _1.0.1: https://github.com/sprockets/sprockets-dynamodb/compare/1.0.0...1.0.1